feat(admin): move sorting & filtering logic from `ChartIndexPage` to `ChartList`
