+++ adminSiteClient/admin.scss
@@ -763,8 +763,7 @@ main:not(.ChartEditorPage):not(.GdocsEditPage) {
     margin-top: 6px;
 }
 
-.ChartIndexPage,
-.UsersIndexPage {
+.ChartList {
     .topRow {
         display: flex;
         align-items: center;
